WIP
This is more difficult than expected. The current issue is the FTP
upload. FTP expects to dial back to us for the data transfer, which
doesn't work because we're behind a NAT inside the VM. I couldn't get
passive mode to work with the ftp lib we're using. So fuck it for now.